// src/types/woocommerce.d.ts
export interface Product {
	id: number;
	name: string;
	slug: string;
	permalink: string;
	date_created: Date;
	date_created_gmt: Date;
	date_modified: Date;
	date_modified_gmt: Date;
	type: "simple" | "variable" | "grouped" | "external";
	status: "any" | "draft" | "pending" | "private" | "publish";
	featured: boolean;
	catalog_visibility: "visible" | "catalog" | "search" | "hidden";
	description: string;
	short_description: string;
	sku: string;
	price: string;
	regular_price: string;
	sale_price: string;
	date_on_sale_from: Date | null;
	date_on_sale_from_gmt: Date | null;
	date_on_sale_to: Date | null;
	date_on_sale_to_gmt: Date | null;
	price_html: string;
	on_sale: boolean;
	purchasable: boolean;
	total_sales: number;
	virtual: boolean;
	downloadable: boolean;
	downloads: any[]; // TODO look at Downloads properties
	download_limit: number;
	download_expiry: number;
	external_url: string;
	button_text: string;
	tax_status: "taxable" | "shipping" | "none";
	tax_class: "standard" | "reduced-rate" | "zero-rate";
	manage_stock: boolean;
	stock_quantity: number;
	stock_status: "instock" | "outofstock" | "onbackorder";
	backorders: "no" | "notify" | "yes";
	backorders_allowed: boolean;
	backordered: boolean;
	sold_individually: boolean;
	weight: string;
	dimensions: Dimensions;
	shipping_required: boolean;
	shipping_taxable: boolean;
	shipping_class: string;
	shipping_class_id: number;
	reviews_allowed: boolean;
	average_rating: string;
	rating_count: number;
	related_ids: number[];
	upsell_ids: number[];
	cross_sell_ids: number[];
	cross_sell_products?: Product[]
	parent_id: number;
	purchase_note: string;
	categories: Partial<Category>[];
	tags: any[]; // TODO look at Tags properties
	images: Image[];
	attributes: Attribute[];
	default_attributes: any[]; // TODO look at default attributes properties
	variations: number[] | Variation[];
	grouped_products: number[];
	menu_order: number;
	meta_data: MetaDatum[];
	_links: Links;
	acf: any;
}

export interface Category {
	id: number;
	name: string;
	slug: string;
	parent: number;
	description: string;
	display: string;
	image: Image;
	menu_order: number;
	count: number;
	_links: Links;
}

export interface Collection {
	href: string;
}

export interface Dimensions {
	length: string;
	width: string;
	height: string;
}

export interface MetaDatum {
	id: number;
	key: string;
	value: string;
}
// ...
interface Attribute {
	id: number;
	name: string;
	position: number;
	visible: boolean;
	variation: boolean;
	options: string[];
}

export interface Image {
	id: number;
	date_created: Date;
	date_created_gmt: Date;
	date_modified: Date;
	date_modified_gmt: Date;
	src: string;
	name: string;
	alt: string;
	position: number;
}

export interface Attribute {
	id: number;
	name: string;
	option: string;
}

export interface MetaData {
	id: number;
	key: string;
	value: string;
}

export interface Up {
	href: string;
}
// ...
export interface Links {
	self: Self[];
	collection: Collection[];
	up: Up[];
}
// ...
export interface Variation {
	id: number;
	date_created: Date;
	date_created_gmt: Date;
	date_modified: Date;
	date_modified_gmt: Date;
	description: string;
	permalink: string;
	sku: string;
	price: string;
	regular_price: string;
	sale_price: string;
	date_on_sale_from?: any;
	date_on_sale_from_gmt?: any;
	date_on_sale_to?: any;
	date_on_sale_to_gmt?: any;
	on_sale: boolean;
	visible: boolean;
	purchasable: boolean;
	virtual: boolean;
	downloadable: boolean;
	downloads: any[];
	download_limit: number;
	download_expiry: number;
	tax_status: string;
	tax_class: string;
	manage_stock: boolean;
	stock_quantity?: any;
	in_stock: boolean;
	backorders: string;
	backorders_allowed: boolean;
	backordered: boolean;
	weight: string;
	dimensions: Dimensions;
	shipping_class: string;
	shipping_class_id: number;
	image: Image;
	attributes: Attribute[];
	menu_order: number;
	meta_data: MetaData[];
	_links: Links;
}

export interface Self {
	href: string;
}